Converting everything to int32s.
jsonpb wraps int64s as a string which we dont like.  We'll figure out how to use values larger than int64s when it comes up.

+++ internal/servers/controller/handlers/errors.go
@@ -34,7 +34,7 @@ func InvalidArgumentErrorf(msg string, fields []string) error {
 
 func statusErrorToApiError(s *status.Status) *pb.Error {
 	apiErr := &pb.Error{}
-	apiErr.Status = int64(runtime.HTTPStatusFromCode(s.Code()))
+	apiErr.Status = int32(runtime.HTTPStatusFromCode(s.Code()))
 	apiErr.Message = s.Message()
 	apiErr.Code = s.Code().String()
